+++ controller.go
@@ -28,7 +28,7 @@ import (
 	"strings"
 
 	"github.com/astaxie/beego/context"
-	"github.com/astaxie/beego/param"
+	"github.com/astaxie/beego/context/param"
 	"github.com/astaxie/beego/session"
 )
