+++ parser.go
@@ -30,8 +30,8 @@ import (
 	"strings"
 	"unicode"
 
+	"github.com/astaxie/beego/context/param"
 	"github.com/astaxie/beego/logs"
-	"github.com/astaxie/beego/param"
 	"github.com/astaxie/beego/utils"
 )
 
@@ -39,7 +39,7 @@ var globalRouterTemplate = `package routers
 
 import (
 	"github.com/astaxie/beego"
-	"github.com/astaxie/beego/param"
+	"github.com/astaxie/beego/context/param"
 )
 
 func init() {
